Move Build subcommand to its own module.
So it's easier to add new commands.
